Reorganize tests for __main__
